+++ app/src/main/java/com/github/llmaximll/gym/dataclasses/Exercise.kt
@@ -4,6 +4,6 @@ class Exercise {
     var name: String = "name"
     var numberEx: String = "0"
     var scores: Int = 0
-    var minutes: Int = 0
+    var minutes: Long = 0
     var cal: Float = 0f
 }
